+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Make all render settings available to the CLI
 * Make CLI errors from invalid arguments log more useful messages to
   stderr and give exit code 1.
+* Support trailing commas in inline choices
 
 ### 0.0.22:
 * Remove accidentally-left-in debug log on regex matcher parsing