+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 0.0.23-dev:
+### 0.0.23:
 * Make all render settings available to the CLI
 * Make CLI errors from invalid arguments log more useful messages to
   stderr and give exit code 1.